+++ classes/Request/InvalidJSONRequestException.php
@@ -1,0 +1,6 @@
+<?php
+
+class Request_InvalidJSONRequestException extends UnexpectedValueException
+{
+
+}