+++ CHANGELOG.md
@@ -2,6 +2,7 @@ You're really going to want to read this.
 
 ## Unreleased
 
+* Add `Request::jsonBody` and `Request::jsonBodyArray` for easy access to JSON request bodies.
 * Add `Response::setJSON` and `Controller::respondJSON` as helpers for sending JSON responses.
 
 ## 4.10.0 (2022-11-09)